requirements/requirements.txt
anyio==3.6.2
asyncpg==0.27.0
psycopg2-binary==2.9.5
fastapi==0.85.2
pydantic==1.10.2
SQLAlchemy[asyncio]==2.0.15
uvicorn==0.19.0
loguru==0.6.0
python-dotenv==0.21.0
alembic==1.8.1